Added "DamageFormula" which contains some helpful utility functions for calculating damage. Combattants now have a level, which plays into certain damage calculations. The example skills in Skills.scala now use formulas from DamageFormula to calculate their damage outputs.

+++ src/main/scala/io/dylemma/battle/DamageType.scala
@@ -12,7 +12,7 @@ sealed trait DamageType {
 }
 
 /** Damage, as represented by an amount and a type */
-case class Damage(amount: Int, kind: DamageType)
+case class Damage(amount: Double, kind: DamageType)
 
 object DamageType {
