+++ src/main/scala/io/dylemma/battle/ResourceModificationProcessor.scala
@@ -12,7 +12,7 @@ class ResourceModificationProcessor extends EventHandler {
 	def handlePostEvent(mods: BattleModifiers) = {
 		case DamageResource(target, res, dmg) =>
 			val resource = target getResource res
-			val depleted = resource deplete dmg.amount
+			val depleted = resource deplete dmg.amount.toInt
 			debug(s"$target lost $depleted $res")
 			Nil // TODO - death conditions?
